pyproject.toml
[project]
name = "framework-multiagentes"
version = "0.1.0"
description = "Plataforma reutilizável de agentes de IA com guardrails, observabilidade e avaliação"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langfuse>=2.50.0",
    "deepeval>=2.0.0",
    "chainlit>=2.0.0",
    "qdrant-client>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "tavily-python>=0.5.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentkit"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["_references", ".claude"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "N", "SIM", "RET"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"**/test_*.py" = ["S101"]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
exclude = ["_references/", ".claude/"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["-ra", "--strict-markers", "--strict-config"]
filterwarnings = ["ignore::DeprecationWarning"]
asyncio_mode = "auto"
